[#562] add disclaimer modal to oligrapher page
- remove redundant MapsController spec file
