+++ app/views/maps/_disclaimer_modal.html.erb
@@ -1,0 +1,26 @@
+<div class="modal fade" tabindex="-1" role="dialog" id="oligrapher-disclaimer">
+  <div class="modal-dialog" role="document">
+    <div class="modal-content">
+      <div class="modal-header">
+        <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+        <h4 class="modal-title">Disclaimer</h4>
+      </div>
+      <div class="modal-body">
+        <p><strong>User-contributed maps don't represent the views of LittleSis</strong></p>
+	<br />
+	<p><%= link_to "Click here", "https://littlesis.org/disclaimer", "_target" => 'blank'  %> to read our full disclaimer.</p>
+	
+      </div>
+    </div><!-- /.modal-content -->
+  </div><!-- /.modal-dialog -->
+</div><!-- /.modal -->
+
+
+<%= javascript_tag do %>
+ $(function(){
+   $("a[href='#disclaimer']")
+     .click(function() {
+       $('#oligrapher-disclaimer').modal('toggle');
+     })
+ });
+<% end %>
