+++ app/views/maps/story_map.html.erb
@@ -30,6 +30,8 @@
     </div>
 </div>
 
+<%= render partial: 'disclaimer_modal' %>
+
 <div id="embed-code">
     Width <input type="text" id="embed-width" value="600">&nbsp;&nbsp;
     Height <input type="text" id="embed-height" value="700">&nbsp;&nbsp;
